+++ src/main.cpp
@@ -9,8 +9,8 @@ int main(int argc, char *argv[])
 {
     QApplication app(argc, argv);
     QLocale::setDefault(QLocale(QLocale::Arabic, QLocale::SaudiArabia));
-    QApplication::setApplicationName(QStringLiteral("Arabic Code Studio"));
-    QApplication::setOrganizationName(QStringLiteral("Arabic Code Studio"));
+    QApplication::setApplicationName(QStringLiteral("Lisan Studio"));
+    QApplication::setOrganizationName(QStringLiteral("Lisan Studio"));
 
     MainWindow window;
     window.show();